SL-83: Add dynamic timeout shortening on offline servers
On servers that the poller has marked as offline, shorten the request
timeouts. We're not really expecting the request to succeed, and it's
important to not tie up a thread for too long on a request that we think
is going to fail.

If it does go through, tho, that's a nice bonus.